data_ingestion/requirements.txt
azure-core==1.29.2
azure-eventhub==5.11.4
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
newsapi-python==0.2.7
requests==2.31.0
six==1.16.0
typing_extensions==4.7.1
urllib3==2.0.4
